Add E2E test for AKV key provider
